refactor(022/extract-common): create extractors/ dir, move types + cross-sibling helpers to common.rs
Phase 1 of milestone 022 — convert mikebom-cli/src/parity/extractors.rs
(single 1654-LOC file) to mikebom-cli/src/parity/extractors/{mod,common}.rs
(directory module, history-preserving rename via `git mv`).

extractors/common.rs (NEW, 378 LOC) holds the cross-sibling surface:
- pub struct ParityExtractor                  (cross-format type)
- pub enum Directionality                     (cross-format enum)
- pub fn extract_mikebom_annotation_values    (used by SPDX 2/3 macros)
- extract_spdx23_annotation_values            (private)
- extract_spdx3_annotation_values             (private)
- pub(super) fn decode_envelope               (used by mod.rs C-section
                                               extractors temporarily —
                                               moves into spdx2/spdx3
                                               in commits 3+4)
- pub(super) fn canonicalize_atomic_values    (used by CDX + SPDX paths)
- pub fn walk_cdx_components                  (CDX walker, kept here so
                                               spdx_relationship_edges
                                               can reach walk_spdx23_packages
                                               without an upward dep)
- pub fn walk_spdx23_packages
- pub fn walk_spdx3_packages
- pub(super) fn empty                         (collapsed with the old
                                               g_empty — same body,
                                               two definitions in pre-022)
- pub(super) fn normalize_alg                 (used by all 3 hash extractors)
- pub(super) fn spdx_relationship_edges       (169 LOC, shared SPDX 2/3
                                               graph traversal — biggest
                                               cross-sibling helper)

extractors/mod.rs (1317 LOC, was 1654) now:
- declares `mod common;`
- re-exports the 6 public items from common via `pub use common::{...};`
  preserving the public API path `mikebom::parity::extractors::*`
- holds the format-specific extractors temporarily — these move out
  in commits 2-4 (cdx, spdx2, spdx3)
- holds the EXTRACTORS table (unchanged) and the 2 structural tests

External callers untouched: `mikebom-cli/src/cli/parity_cmd.rs` and
`mikebom-cli/tests/holistic_parity.rs` see identical paths.

Visibility ladder per FR-003:
- pub for items in the documented public surface (re-exported via mod.rs)
- pub(super) for items used by sibling format submodules
- private for items used only within common itself
